Prevent old nuke scripts from running in production (#13943)
## Summary of the problem

Nothing stops these very dangerous scripts from running in production.


## Describe your changes

This PR prevents these scripts from running outside of development.

+++ app/services/pending_event_mapping_engine/nuke.rb
@@ -3,6 +3,8 @@
 module PendingEventMappingEngine
   class Nuke
     def run
+      return unless Rails.env.development?
+
       CanonicalPendingEventMapping.delete_all
     end
