+++ app/services/pending_transaction_engine/nuke.rb
@@ -3,6 +3,8 @@
 module PendingTransactionEngine
   class Nuke
     def run
+      return unless Rails.env.development?
+
       CanonicalPendingEventMapping.delete_all
       CanonicalPendingTransaction.delete_all
       RawPendingStripeTransaction.delete_all
