+++ app/services/transaction_engine/nuke.rb
@@ -3,6 +3,8 @@
 module TransactionEngine
   class Nuke
     def run
+      return unless Rails.env.development?
+
       Fee.delete_all
       CanonicalEventMapping.delete_all
